Add dbc member class, refactor other classes
