+++ apprentice_teacher.rb
@@ -1,6 +1,8 @@
-class ApprenticeTeacher
-  attr_reader :age, :salary, :phase, :target_raise
-  attr_accessor :name
+require_relative 'teacher'
+
+class ApprenticeTeacher < Teacher
+  attr_reader :age, :salary, :phase
+  attr_accessor :name, :target_raise
 
   def initialize(options={})
     @age = options.fetch(:age, 0)
@@ -9,31 +11,13 @@ class ApprenticeTeacher
     @phase = 3
   end
 
-  def offer_high_five
-    "High five!"
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
-
-  def teach_stuff
-    response = ""
-    response += "Listen, class, this is how everything works. "
-    response += "*drops crazy knowledge bomb* "
-    response += "... You're welcome."
-    response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
+   def teach_stuff
+     response = ""
+     response += "Listen, class, this is how everything works. "
+     response += "*drops crazy knowledge bomb* "
+     response += "... You're welcome."
+     response
+   end
 
   def set_performance_rating(rating)
     response = ""
