+++ dbc_member.rb
@@ -1,0 +1,32 @@
+class DbcMember
+  attr_reader :age, :salary, :phase, :performance_rating, :target_raise
+  attr_accessor :name
+
+  def initialize(options={})
+    @phase = 3
+    @age = options.fetch(:age, 0)
+    @name = options.fetch(:name, "")
+    @target_raise = 1000
+  end
+
+  def offer_high_five
+    "High five!"
+  end
+
+  def set_phase(num)
+    @phase = num
+    "Cool, I've always wanted to teach phase #{num}!"
+  end
+
+  def salary=(new_salary)
+    puts "This better be good!"
+    @salary = new_salary
+  end
+
+  def learn_stuff
+    response = ""
+    response += "WHOA! I've never thought of it quite like that before. "
+    response += "Now I feel like a genius!"
+    response
+  end
+end
