+++ student.rb
@@ -1,4 +1,6 @@
-class Student
+require_relative 'dbc_member'
+
+class Student < DbcMember
   attr_reader :age, :phase
   attr_accessor :name
 
@@ -6,10 +8,6 @@ class Student
     @phase = 1
     @age = options.fetch(:age, 0)
     @name = options.fetch(:name, "")
-  end
-
-  def offer_high_five
-    "High five!"
   end
 
   def set_phase(num)
@@ -24,10 +22,4 @@ class Student
     response
   end
 
-  def learn_stuff
-    response = ""
-    response += "WHOA! I've never thought of it quite like that before. "
-    response += "Now I feel like a genius!"
-    response
-  end
 end
