+++ teacher.rb
@@ -1,22 +1,8 @@
-class Teacher
+require_relative 'dbc_member'
+
+class Teacher < DbcMember
   attr_reader :age, :salary, :phase, :performance_rating, :target_raise
   attr_accessor :name
-
-  def initialize(options={})
-    @phase = 3
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-    @target_raise = 1000
-  end
-
-  def offer_high_five
-    "High five!"
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
 
   def teach_stuff
     response = ""
@@ -24,15 +10,6 @@ class Teacher
     response += "*drops flat-out insane knowledge bomb* "
     response += "... You're welcome. *saunters away*"
     response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
   end
 
   def set_performance_rating(rating)
@@ -45,5 +22,9 @@ class Teacher
       response += "feedback, I'll do better next time."
     end
     response
+  end
+
+  def receive_raise(raise)
+    @salary += raise
   end
 end
